move listing method to DockmasterCommand trait
